Factor out super class functions of contract-loader
