+++ packages/colony-js-contract-loader/defaults.js
