+++ packages/colony-js-contract-loader/package.json
@@ -38,5 +38,8 @@
     "engines": {
         "node": ">=8.2.1",
         "yarn": ">=1.3.2"
+    },
+    "dependencies": {
+        "assert": "^1.4.1"
     }
 }
